requirements.txt
langchain
python-dotenv
ipykernel
langchain-google-genai
langchain_community
langchain_core
streamlit
google-generativeai